create product

+++ src/main/java/com/vanndeth/phoneshopmgt/repository/ModelRepository.java
@@ -9,8 +9,8 @@ import org.springframework.stereotype.Repository;
 import com.vanndeth.phoneshopmgt.entity.Model;
 
 @Repository
-public interface ModelRepository extends JpaRepository<Model, Integer>{
+public interface ModelRepository extends JpaRepository<Model, Long>{
 	
-	List<Model> findModelByBrandId(Integer brandId);
+	List<Model> findModelByBrandId(Long brandId);
 
 }
